requirements.txt
duckduckgo-search
faster-whisper
google-genai
groq
numpy
psutil
pygame
python-dotenv
sentence-transformers
silero-vad
sounddevice
soundfile
torch
webrtcvad
